requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
pytest==7.4.3
black==23.10.0
python-telegram-bot==20.7
APScheduler==3.10.4